requirements.txt
cassandra-driver
pyspark
kafka-python
requests
apache-airflow
pandas
ta
PyArrow
ccxt
backoff